Split objects.rb file into two files: board.rb and player.rb
