+++ bin/main.rb
@@ -1,6 +1,7 @@
 #!/usr/bin/env ruby
 
-require_relative '../lib/objects'
+require_relative '../lib/board'
+require_relative '../lib/player'
 require_relative '../lib/tictactoe'
 
 board = Board.new
